+++ src/main.lua
@@ -3,6 +3,8 @@
 -- these funky (---@) comments are just there
 --	 to help VS Code find the definitions of things
 
+import = require
+
 ---@diagnostic disable-next-line: undefined-global
 local mods = rom.mods
 
@@ -49,16 +51,35 @@ local function on_ready()
 	import "Scripts/CharonsTipJar.lua"
 end
 
+-- Loaded after all other mods
+-- Define static Context Wraps in here to prevent issues as per https://github.com/SGG-Modding/ModUtil/issues/12
+local function on_ready_late()
+	if config.enabled == false then return end
+
+	import "Scripts/RunClearData_Late.lua"
+end
+
 local function on_reload()
 	-- what to do when we are ready, but also again on every reload.
 	-- only do things that are safe to run over and over.
 	if config.enabled == false then return end
 end
 
+local function on_reload_late()
+	if config.enabled == false then return end
+end
+
 -- this allows us to limit certain functions to not be reloaded.
-local loader = reload.auto_single()
+local loader = reload.auto_multiple()
 
 -- this runs only when modutil and the game's lua is ready
 modutil.once_loaded.game(function()
-	loader.load(on_ready, on_reload)
+	loader.load("early", on_ready, on_reload)
+end)
+
+-- again but loaded later than other mods
+mods.on_all_mods_loaded(function()
+	modutil.once_loaded.game(function()
+		loader.load("late", on_ready_late, on_reload_late)
+	end)
 end)